move reconstruct js response to L8ResponseObject::reconstruct_js_response
